+++ Samples~/Demo/Scripts/SDKClient.cs
@@ -1,8 +1,7 @@
 using Nmkr.Sdk.Schemas;
-using UnityEngine;
 using static Nmkr.Sdk.Api;
 using System.Threading.Tasks;
-using static Nmkr.Demo.SDKServer;
+using Nmkr.Sdk;
 
 namespace Nmkr.Demo
 {
@@ -11,7 +10,7 @@ namespace Nmkr.Demo
     /// Use multiplayer solutions like RPC calls.
     /// !DONT USE THIS DEMO IN PRODUCTION. YOU ARE RESPONSIBLE FOR THE SECURITY OF YOUR API KEYS AND OTHER SENSITIVE DATA!
     /// </summary>
-    public class SDKClient : MonoBehaviour
+    public class SDKClient
     {
         public static async Task<ApiResponse<WalletInfo>> CreateWallet(string walletName, string walletPassword)
         {
@@ -28,33 +27,9 @@ namespace Nmkr.Demo
             return await SDKServer.GetWalletAssets(address);
         }
 
-        public static async Task<ApiResponse<NftProjectsDetails>> GetProjectDetails(string projectUid)
-        {
-            return await SDKServer.GetProjectDetails(projectUid);
-        }
-
         public static async Task<ApiResponse<MakeTransactionResultClass>> MakeManagedWalletTransaction(string senderaddress, string walletpassword, string receiverAddress, TransactionTokensClass token)
         {
-            var transactionRequest = new CreateManagedWalletTransactionClass()
-            {
-                receivers = new TransactionReceiversClass[]
-                {
-                    new TransactionReceiversClass()
-                    {
-                        receiverAddress = receiverAddress,
-                        sendTokens = new TransactionTokensClass[]
-                        {
-                            token
-                        }
-                    },
-                }
-            };
-            return await SDKServer.MakeManagedWalletTransaction(senderaddress, walletpassword, transactionRequest);
-        }
-
-        public static async Task<ApiResponse<MintAndSendResultClass>> MintAndSendRandom(string projectUid, int nftCount, string receiverAddress)
-        {
-            return await SDKServer.MintAndSendRandom(projectUid, nftCount, receiverAddress);
+            return await SdkWrapper.MakeManagedWalletTransaction(senderaddress, walletpassword, receiverAddress, token);
         }
 
         public static async Task<ApiResponse<PurchaseAndMintResult>> PurchaseAndMint(string projectUid, string managedWalletAddress, string managedWalletPassword, long itemPriceLovelace, int nftCount)