+++ Samples~/Demo/Scripts/SDKServer.cs
@@ -1,55 +1,20 @@
 using Nmkr.Sdk;
 using Nmkr.Sdk.Schemas;
-using System;
-using System.Security.Policy;
-using System.Threading;
 using System.Threading.Tasks;
-using UnityEditor.PackageManager.Requests;
-using static Nmkr.Demo.SDKServer;
 using static Nmkr.Sdk.Api;
 
 namespace Nmkr.Demo
 {
+    /// <summary>
+    /// This class is a starter point for communication between a Unity game client and the Unity game server.
+    /// Use multiplayer solutions like RPC calls.
+    /// !DONT USE THIS DEMO IN PRODUCTION. YOU ARE RESPONSIBLE FOR THE SECURITY OF YOUR API KEYS AND OTHER SENSITIVE DATA!
+    /// </summary>
     public class SDKServer
     {
-        /// <summary>
-        /// UNITY_SERVER define strips this code from non Server Builds. To activate this define choose Dedicated Server in Build Settings.
-        /// This is to prevent sensitive data from being exposed to the client. 
-        /// !DONT USE THIS DEMO IN PRODUCTION. YOU ARE RESPONSIBLE FOR THE SECURITY OF YOUR API KEYS AND OTHER SENSITIVE DATA!
-        /// </summary>
-#if UNITY_SERVER || UNITY_EDITOR
-        public void InitializeApi(string apiKey, int customerId)
-        {
-            ApiSettings settings = new ApiSettings(server: ApiServer.Preprod, apiKey, customerId);
-            Initialize(settings);
-        }
-#endif
-
         public static async Task<ApiResponse<WalletInfo>> CreateWallet(string walletName, string walletPassword)
         {
-            var response = await Api.CreateWallet(walletpassword: walletPassword, enterpriseaddress: true, walletname: walletName);
-
-            var createWalletResponse = new ApiResponse<WalletInfo>();
-
-            if (response.success)
-            {
-                var walletResult = response.result;
-                var wallet = new WalletInfo()
-                {
-                    walletName = walletResult.walletName,
-                    address = walletResult.address,
-                    addressType = walletResult.addressType,
-                    network = walletResult.network,
-                };
-                createWalletResponse.result = wallet;
-                return createWalletResponse;
-            }
-            else
-            {
-                createWalletResponse.error = response.error;
-                return createWalletResponse;
-            }
-
+            return await SdkWrapper.CreateWallet(walletName, walletPassword);
         }
 
         public static async Task<ApiResponse<TxInAddressesClass>> GetWalletUtxo(string walletAddress)
@@ -59,24 +24,7 @@ namespace Nmkr.Demo
 
         public static async Task<ApiResponse<WalletAssets>> GetWalletAssets(string walletAddress)
         {
-            var response = await Api.GetWalletUtxo(walletAddress);
-            var resultResponse = new ApiResponse<WalletAssets>();
-
-            if (response.success)
-            {
-                var assets = new WalletAssets()
-                {
-                    balanceInLovelace = response.result.lovelaceSummary,
-                    tokens = response.result.getAllTokens
-                };
-                resultResponse.result = assets;
-                return resultResponse;
-            }
-            else
-            {
-                resultResponse.error = response.error;
-                return resultResponse;
-            }
+            return await SdkWrapper.GetWalletAssets(walletAddress);
         }
 
         public static async Task<ApiResponse<NftProjectsDetails>> GetProjectDetails(string projectUid)
@@ -84,9 +32,9 @@ namespace Nmkr.Demo
             return await Api.GetProjectDetails(projectUid);
         }
 
-        public static async Task<ApiResponse<MakeTransactionResultClass>> MakeManagedWalletTransaction(string senderaddress, string walletpassword, CreateManagedWalletTransactionClass request)
+        public static async Task<ApiResponse<MakeTransactionResultClass>> MakeManagedWalletTransaction(CreateManagedWalletTransactionClass request)
         {
-            return await Api.MakeTransaction(senderaddress, walletpassword, request);
+            return await Api.MakeTransaction(request);
         }
 
         public static async Task<ApiResponse<MintAndSendResultClass>> MintAndSendRandom(string projectUid, int nftCount, string receiverAddress)
@@ -96,73 +44,7 @@ namespace Nmkr.Demo
 
         public static async Task<ApiResponse<PurchaseAndMintResult>> PurchaseAndMint(string projectUid, string managedWalletAddress, string managedWalletPassword, long itemPriceLovelace, int nftCount)
         {
-            var response = new ApiResponse<PurchaseAndMintResult>();
-
-            //Check wallet funds
-            ApiResponse<TxInAddressesClass> getWalletResponse = await GetWalletUtxo(managedWalletAddress);
-            if (getWalletResponse.success)
-            {
-                TxInAddressesClass walletUtxo = getWalletResponse.result;
-                if (walletUtxo.lovelaceSummary < itemPriceLovelace)
-                {
-                    response.error = getWalletResponse.error;
-                    return response;
-                }
-            }
-            else
-            {
-                response.error = getWalletResponse.error;
-                return response;
-            }
-
-            // Get project payout wallet
-            var projectDetailsResponse = await GetProjectDetails(projectUid);
-            if (projectDetailsResponse.success)
-            {
-                NftProjectsDetails projectDetails = projectDetailsResponse.result;
-                string payoutAddress = projectDetails.adaPayoutWalletAddress;
-                // Make transaction
-                var transactionRequest = new CreateManagedWalletTransactionClass()
-                {
-                    receivers = new TransactionReceiversClass[]
-                    {
-                        new TransactionReceiversClass()
-                        {
-                            receiverAddress = payoutAddress,
-                            receiverLovelace = itemPriceLovelace
-                        },
-                    }
-                };
-
-                var transactionResponse = await MakeManagedWalletTransaction(managedWalletAddress, managedWalletPassword, transactionRequest);
-                if (transactionResponse.success && transactionResponse.result.state == MakeTransactionResults.success)
-                {
-                    var mintResponse = await Api.MintAndSendRandom(projectUid, nftCount, receiveraddress: managedWalletAddress);
-                    if (mintResponse.success)
-                    {
-                        response.result = new PurchaseAndMintResult()
-                        {
-                            nfts = mintResponse.result.sendedNft
-                        };
-                        return response;
-                    }
-                    else
-                    {
-                        response.error = transactionResponse.error;
-                        return response;
-                    }
-                }
-                else
-                {
-                    response.error = transactionResponse.error;
-                    return response;
-                }
-            }
-            else
-            {
-                response.error = projectDetailsResponse.error;
-                return response;
-            }
+            return await SdkWrapper.PurchaseAndMint(projectUid, managedWalletAddress, managedWalletPassword, itemPriceLovelace, nftCount);
         }
     }
 }