+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.0-pre.6] - 2024-02-29
+### Added
+* SdkWrapper class
+### Changed
+* SdkWrapper.GetWalletAssets now gets the tokens found in the txIn
+* Changed Demo classes SDKClient and SDKServer to utilize SdkWrapper class
+
 ## [1.0.0-pre.5] - 2024-02-08
 * API Changes and Demo Readme changes
 ### Known Issues